[rf] Move `ApplyClientConnectionConfigurationToRESTConfig` function to `client/kubernetes` package
