+++ extensions/pkg/util/clientset.go
@@ -8,6 +8,8 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 	componentbaseconfig "k8s.io/component-base/config"
+
+	"github.com/gardener/gardener/pkg/client/kubernetes"
 )
 
 // NewRESTConfigFromKubeconfig creates a new REST config from a given Kubeconfig and returns it.
@@ -23,16 +25,7 @@ func NewRESTConfigFromKubeconfig(kubeconfig []byte) (*rest.Config, error) {
 
 // ApplyClientConnectionConfigurationToRESTConfig applies the given client connection configurations to the given
 // REST config.
-func ApplyClientConnectionConfigurationToRESTConfig(clientConnection *componentbaseconfig.ClientConnectionConfiguration, rest *rest.Config) {
-	if clientConnection == nil {
-		return
-	}
-
-	rest.AcceptContentTypes = clientConnection.AcceptContentTypes
-	rest.ContentType = clientConnection.ContentType
-	rest.Burst = int(clientConnection.Burst)
-	rest.QPS = clientConnection.QPS
-}
+var ApplyClientConnectionConfigurationToRESTConfig = kubernetes.ApplyClientConnectionConfigurationToRESTConfig
 
 // createRESTConfig creates a Config object for a rest client. If a clientConnection configuration object is passed
 // as well then the specified fields will be taken over as well.
